+++ scripts/generate-sitemap.mjs
@@ -257,6 +257,7 @@ ${urls
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9"
         xmlns:image="http://www.google.com/schemas/sitemap-image/1.1">
 ${products
+  .filter((p) => !GOOGLE_SHOPPING_BLOCKED.has(p.id))
   .map(
     (p) => `  <url>
     <loc>${SITE_URL}/products/${p.slug}</loc>